Using marker files to designate a package for a Guild op
Certain packages must be available on the Python path for an
operation. These include guild itself and psutil (for the various sys
plugins). Others may follow.

These packages must be marked with a '__guild_op_package__' file in
their root. See third-party/psutil.BUILD for how a genrule can be used
for this.