Cleanup op references
An op reference is how we associate a run with its operation. It's
stored as a run attribute `opref` and contains information about the
model and the op name.

`op.py` now officially handles all run attrs including op
references. It manges the encoding and decoding of the reference via
the `OpRef` class.

This change also introduces a scheme where a run can be initialized as
a part of an iteration in `var`. `runs_impl` uses this to init an op
ref from the run attr and associate it with the run.