Run op_main as a run script rather than as a module
We run into problems with imports when we run as a module. The module
name is "__main__", which shift the top-level module to `guild`
([reference](https://www.python.org/dev/peps/pep-0328/#id11)). Any
imports that happen to be named with a module under `guild` will be
shadowed by the Guild module. Running `op_main` from a script avoids
this problem.