Fix bug in cmd spec parsing
For the record, cmd specs are specified as:

    [PACKAGE_PATH + '/'] MODULE_PATH

`PACKAGE_PATH` is a `/` delimted path pointing to the package
root. This will be inserted into the system path to make the module
available.

`MODULE_PATH` is a `.` delimited path pointing to a module that can be
found on the system path.