Changes to resource support
- Operation sources no longer have paths - use select instead
- Prefix is changed to select (see below)
- Operation sources may include a list of operations (see below)

The select attribute of a source is used to select a subpath of the
resource source. This applies to source directories, archives, and
operation sources.

The select attribute is a regular expression that is matched against
the whole string (i.e. implicitly is wrapped in ^...$). Matching paths
are linked to the run directory using their base name.

For example, the select value `data` will match only a `data` path in
the source. To select all the files under `data`, use `data/.*`.

Selected paths may include a group, which is used as an alternative
matching path.

Operation sources may include a list of operations. The latest of any
matching operation is used.