Fix broken build on macos
The primary issues is that tensorflow 1.6.0 now requires grpcio, which
doesn't have a wheel for macos, at least for its latest version. We're
explicitly installing a version of grpcio that does to work around
this.

This is being applied to all targeted operation systems.

Commit also includes removal of "quiet" features, which ended up
burying useful information.