Formally define 'location' attribute for a resource
The resource location is used by source resolvers to find a source
relative to the resource. This is really used by the file resolvers,
as url and operation sources are independent of the defined resource.

For file sources defined in a guildfile, the location is the guildfile
directory. For file sources defined in a package, the location is the
package directory under the distribution.