Revamp resource download/caching logic
We were relying on cachecontrol from pip and when that went away it
revealed a bug with caching.

Now that we're not using a content cache, we don't want to be so
aggressive in re-downloading invalid files (hash mismatch). As these
files can be costly to download, a mismatch is handled by simply
stopping with an error and not re-downloading.

Users will be responsible for re-downloading bad files as needed.