Use "0.0.0.0" for default host if none is specified.
Thanks to @cfregly to pointing out that '0.0.0.0' is the correct
default, rather than an empty string ('').

See https://docs.python.org/2/library/socket.html#socket.getaddrinfo
for details on the function. While in other parts of the docs, it
states that "the empty string represents INADDR_ANY", an empty string
does not represent all interfaces for `getaddrinfo`, which is what
werkzeug uses to setup its socket.