Support for operation pre-process script
This commit introduces a hook for running a shell script command after
operation dependencies are resolved and before the operation itself.

Prior to this feature, pre-processing would have to be run in the
operation itself. While pre-processing in the operation is still a
viable option, the new functionality allows model devs to separate
concerns and use shell script commands.