Remove cwd filter for model related ops
Prior to this commit, Guild would behave differently if the cwd
contained a Guild file. It would limit certain query commands to show
only content associated with models defined in the cwd. For example,
`guild models` would only show those models defined in the cwd Guild
file.

The thinking here was that users who are working on a project will run
commands in the project directory and will want those commands to see
only the project content.

In practice, this got confusing. There are cases where the user will
want to see models and resources defined outside the cwd - for example
when running required operations from an installed package. This is
common enough to make the cwd limit scheme untenable.

This commit removes the limit and adds modifies some commands to
better support queries on the cwd.

For example, now by default, `guild models` will list all visible
models - those installed in the env and those defined in the
cwd. However, `guild models .` (note the dot argument) can be used to
filter the models, limiting them to the cwd.

This feature has been added to operations and resources listings.

There are some areas currently (as of this commit) that may need
similar filtering features. These include `runs list` and `compare`.

Additionally, this commit includes auto-limits to runs list, as now
all runs in the env are listed by default. The auto-limit will display
only the first 20 runs that match the specified filter criteria. All
runs can be listed using `-a, --all` and additional increments of 20
runs can be displayed using one or more `-m, --more` options.