Support for resource-src function in resource scripts
Usage:

    $(resource-src PATH)

This function is resolves by Guild before running the resource
script. It looks for `PATH` in the resource def location (defining
Guild file and parent locations if the resource is defined in an
inherited model).

This establishes a pattern of providing Guild resolved functions in
the form `$(FUNCTION [ARG...])`.