Cleanup guildfile support
- Remove support for model flags
- Remove package resources
- Fix $include occuring in package models

Flags at the model level was an early attempt to support flag reuse
across operations. This is no longer useful as $include can be used
for flag reuse.

Package level resources are also no longer useful as it's possible to
define resources using config mixins or $include.

Prior to this commit, it was not possible to extend package models
that used $include in their Guild files.