Remove matplotlib plugin
What this plugin does--use Agg backend to avoid loading tkinter--is
trivial to implement in a wrapping script if need be. The driving
example was the TensorFlow fashion mnist example, where this issue
pops up.

It might be a good idea to reinstate this if the problem comes up
enough. But as it is, the added 100ms time to start an op, which is
incurred in all cases where matplotlib is installed, is not obviously
worth the payoff.

In addition, it's not at all clear that this behavior is desirable. If
an env needs tkinter, the user can install it.