Revamp init to no longer automatically install TensorFlow
TensorFlow can be installed along with any other required package. We
do still however provide support for installing the GPU enabled
package if the system supports it. This is done by requiring the
special package `tensorflow-any` in the Guild file package def. This
is the only way to get this behavior.

Users can alternatively install TensorFlow using `guild init -r
tensorflow{-gpu}` if the package doesn't explicitly require it.

This change also includes special support for auto-selecting the
Python interpreter on init. A requirement for `python` with a
particular version will tell Guild what default interpreter to use on
init if one is not specified.