Suppor for skipping a run if a matching run is available
The --needed option causes run to first check if there is a non-error
run available that matches the op and flag values. If a matching run
exists, the run is skipped.

There's currenty no way to change the matching logic - it's always a
full match on the op and flag values.

A possible future --needed-flags FLAGS alternative might be useful (as
requested) that uses the flags list in the match, rather than using
all flags.