Bug when TF events saved in run dir root
If a TF event file was saved in the root, its values were replaced by
any values under .guild (output scalars).

Problem was that key prefix was being stored as '' in both cases.

Change was to store the correct prefix ('.guild' in the case of output
scalar location) and deal with stripping '.guild' in presentation.