Support user config for pattern types
Pattern types are "text", "binary", or "dir" and may be used to
qualify a source code pattern for selection.

For example:

```
sourcecode:
  - exclude: {dir: '*'}
```

indicates that the pattern applies only to directories and not to
files. Similarly:

``` sourcecode
  - exclude: {binary: '*'}
```

indicates that only binary files should be excluded, and not text
files or directories.

To help debug selection issues, the `--test-sourcecode` option of
`run` now prints the rules that are applied to files.