Overhaul of flag value decoding/encoding
Gist is that the functions of decoding, encoding, and formatting flag
values have been moved into `guild.flag_util`. New tests (in
`flag-util` and `flat-vals` demonstrate the new functionality).

The driver for this change is #50. Guild was not handling quoted
strings that could be converted to floats as strings - but as floats.