Decent implementation of hparam support
We have to concede to TensorBoard's approach to experiments, which
requires that they be fully specified when created. Experiments cannot
be modified once logged. This means that we have to construct an
experiment using as much information as we have when TensorBoard is
started. New runs are limited to the flags and scalars that are
available from the initial set of runs. Guild TensorBoard has to be
restarted and the browser refreshed for new flags and scalars to be
visible.