Include sourcecode path before modeldef paths
This is a poor man's version of virtual env style isolation. To
properly isolate a run, we should create a package and install it with
dependencies in a virtual env. Additional levels of isolation might
use containers. This is a compromise in the interest of KISS and
running operations quickly.