Refactor the way plugins are enabled
- Plugins must now be explicitly enabled in configuration

- Using a 'provides' scheme where a set of plugins can be
  enabled (e.g. all, summary)

Rationale: Summary plugins are the only operation-time plugins in
effect. These are heavy weight if TensorFlow is installed as they
patch the TF summary writer facility to log additional summary
stats. Just the act of patching these libraries incurs a 1000+ ms
latency to starting the operation. This is a cost that the user should
not bear unless they need summary stats.