Remove pre-process support for operations
Pre-processing let operations perform extra steps to prepare the run
directory before running. To perform these steps, the operation should
now either (a) include the preparation in the operation script, (b)
use a required resource, which may include an operation, to prepare
files, or (c) use a post-process script for a resource source.