Revert output sync
The condition these changes are trying to address is highly
intermittent and shows up in batch output where, on occasion, trial
output is interleaved.

It's noteworthly that the interleaving is by line. Lines are not
mixed.

Going to address this by working around the breaking
test (batch-skopt.md:184 as of this commit)