Special handling for scientitic notation that resembles run IDs
This is driven by https://github.com/guildai/guildai/issues/50

This is a tricky issue. Guild proceses flag values using YAML 1.0
formatting rules but includes some exceptions. One exception is that
it supports float decoding for scientific notation that does not
include periods. So `1e2` is a valid float value in Guild.

The problem arises when a run ID can be parsed as a valid float, as is
highlighted by #50. If Guild parses the value as a float, when it's
used later as a run ID, it will not match. E.g. `1e2` becomes `100.0`.

Guild does not know ahead of time the type of flag value when decoding
values. Only after the value has been decoded will Guild coerce it to
it's target type. This is consistent with loading YAML - values come
in as specified. As flag values are YAML, we can safely apply the
decoding rules ahead of time without knowing the flag type. In
practice this works well. Where a numeric value must be treated as a
string, the user must quote that value.

However, requiring a user to quote what looks like a string (e.g. the
innocuous looking run ID `3e563` from the example) is out of the
question. Therefore Guild now applies a regex to detect these cases
and treats the flag value as a string.

In cases where a float is intended, the user can specify the flag type
as `float` in the Guild file for the operation. See
https://github.com/guildai/issue-resolution/tree/master/50-hash-can-look-like-scientific-notation
for an example.