Attempt to fix issue on macOS with pandas import errors during test
Issue summary:

- Error on macOS Python 3x CI tests - shows up in `imports` test where
  `pandas` is not available - the uat installs pandas, so it's
  surprising that it's not available.

- The issue is that when `guild init` is used to create the uat
  virtual environment, the command to use pip to install guild ends up
  creating a guild shell script with an incorrectly configured shebag
  - it points to the system level Python VM rather than the virtual
  environment Python VM.

- When the same commands to create the uat virtual environment and to
  install the Guild distribution are run from a terminal, they work as
  expected. The issue appears when the exact same commands are run
  from Python using `subprocess.check_call`.

- The difference between running the commands from a terminal and
  running them from Python using `subprocess.check_call` are
  environment related. When the `check_call` function `env` arg is set
  to `{}` (empty dict) the commands from Python works as expected.

This solution is a bit blind - it's not clear what part of the
environment is causing the different behavior. However, the intent of
the Python based commands is to be free from shell specific behavior,
so setting `env` to an empty dict is desireable in any case, assuming
it doesn't break behavior elsewhere.

It's not clear from the [Python API
docs](https://docs.python.org/3/library/subprocess.html) what the
difference between `None` and `{}` are. It does appear from reading
the source of `subproces.py` that when env is None, forkexec uses
execv instead of execve. There may be different behavior related to
`PATH` environment variable when the different commands are used.

In our case, we don't need a path as the explicit path to the Python
exe is sufficient to run the pip command.