Revamp label command
Includes support for appending, prepending, and removing parts of a
label.

**IMPORTANT: This is a CLI beaking change!**

The new interface for setting a label is to use the `--set` or `-s`
option:

    $ guild label -s foo

Prior to this change, the label value was specified as the final
argument. This not only was confusing (does label go first, or last)
it did not support alternative mods of setting the label.

This change introduces a label *action*, which must be specified as
one of: `--set`, `--prepend`, `--append`, `--remove`, or `--clear`.

See `guild/tests/uat/labels.md` for details.