Disabling buffers for operations and API calls
Attempt to resolve test failures where stdout and stderr output is
inconsistent.

The implications of not buffering operation streams _should_ be
minimal, in particular for Python ops, where we run with unbuffered
streams anyway (using -u option). This is nonethess a change that
could potentially introduce behavioral and performance issues for
runs.