Remove --ignore-installed when installing guild dist
This is an experiment to work around bizarre behavior on macOS Python
2.7 where filelock seems to be landing outside the virtual
environment.

```
ERROR: Could not install packages due to an EnvironmentError: [Errno
13] Permission denied: '/usr/local/lib/python2.7/site-packages/filelock-3.0.12.dist-info/RECORD'
Consider using the `--user` option or check the permissions.
```

Short of more pip debugging, this change reverts back to not use
`--ignore-installed`. If we can get away with this (e.g. the pip
behavior in 19.3.1 that prompted the use of `--ignore-installed` is
changed) then not re-installing the reqs has significant performance
gains.

If we run into the issue that prompted this change:

```
    # Install Guild dist ignoring installed. In some cases (behavior
    # seems to have landed in pip 19.3.1) pip will decide not to
    # install the Guild req if it's already installed outside the
    # virtual env. The --ignore-installed flag in this case forces pip
    # to actually install Guild.
    _install_reqs([req], config, ignore_installed=True)
```

then we need to reinstate and deal with the weirdness on macOS.