Fix tests on Linux Python 2.7
See https://circleci.com/gh/guildai/guildai/3407

The beahvior on the system is that the PYTHONPATH ".." is not applied
by Python as the parent directory, but as the current
directory (project/b). This is true when PYTHONPATH is "../.." as
well. This appears to be a bug in the Python interpreter that doesn't
show up on any other platform or version if Python.

Using `projects` here uses the full path, which resolves the issue.