Implement `--proto` run option
`--proto` replaces `--rerun` from 0.6.x.

The term 'rerun' is similar to 'restart' and causes some confusion.

The tern 'prototype' indicates that a run is started using another run
as its prototype. This now includes source code as well as flag
values.

The `--force-sourcecode` may be used to force Guild to use the current
source code when starting a run from a prototype. Otherwise, Guild
uses the prototype source code for the new run.