New run flag: --fail-on-trial-error
Prior to this change, batch ops would fail when a trial fails and
there was no way to tell the trial to continue on on error.

This change breaks that default behavior. Batches continue by default
even when a trial fails.

The new option instructs the batch to fail when a trial fails.

An argument could be made to introduce `--continue-on-trial-error`,
which would preserve the default behavior and use an option to enable
the new behavior explicitly. This follows the "fail fast" principle by
default.

However, machine learning operations may resonably fail due to
outlying hyperparameters, which are explored by batches. To stop the
batch itself in these cases would cause subsequent trials to be
skipped. If the user didn't have the prescience to to use
`--continue-on-trial-error` - not knowing beforehand that
hyperparameters would cause trial errors - a valuable experiment would
be cut short. Therefore, the default cases of continuing is more
conservative than failing fast.

Note that steps, which are not batches (they don't generate trials) do
not continue when a step fails. A stepped operation itself fails as
soon as one of its steps fails. (Arguably, this ought to be
configurable on a step-by-step basis so that any given step can be
designated as "continue on error", however, that topic is not
addressed in this commit.)