Change blocking logic of queues: ignore-running -> wait-for-running
IMPORTANT: This is a breaking change.

Prior to this change, queues waited for other runs to complete before
starting staged runs. The rationale for this behavior is the more
conservative assumption that a system is dedicated to running one
operation at a time.

This default behavior prevents multiple queues from running operations
concurrently. Users must specify `ignore-running` on each queue to
support concurrent queue execution. This is both confusing and
unnecessary. In retrospect, the assumption that a system is dedicated
to one operation at a time can be avoided by simply deferring to the
user: if the system is dedicated to one run at a time, one queue may
be started.

This commit changes the default logic to allow concurrent
operations. A queue may be started with `wait-for-running` to prevent
it from starting staged runs when other runs are in progress.