Use environment Python exe for operations
Prior to this commit, Guild used the Python exe used to run Guild for
operations. This is surprising behavior when a virtual environment is
activated and Guild is installed outside the environment.

The Python exe can be specified explicitly using the environment
variable `GUILD_PYTHON_EXE`.

The order of precedence is:

- `GUILD_PYTHON_EXE` env var
- `CONDA_PYTHON_EXE` env var
- `bin/python` under `VIRTUAL_ENV` if exists
- which python if exists (equivalent to `/bin/usr/env python`)
- Current Python exe (`sys.executable`)