Change choice attr arg-value to alias - different semantic
We can't get away from `value` as being the definitive value for a
flag. There's no provision for maintaining both a value and an 'arg
value' and to do this would complicate the structure used to store
values.

The scheme introduced by this commit is to support a choice 'alias',
which can be used as a short version for a value. The value still
applies, but it can be selected using the specified alias.

Use of alias is optional - the choice value may also be used.