Bug where Guild could not select 'directories' from archives
A 'directory' is a misnomer for an archive - only files are stored and
directories are implied by file paths. Prior to this commit, Guild
didn't recognize the spirit of a directory for archives and would not
select an unpacked directory (which does certainly exist) the way it
does with file resources.