Preliminary support for target-type
target-type can be specified for resources and for sources. Source
level overrides resource level. Not specifying target-type implies the
default.

target-type may be `link` or `copy`. A value of `link` causes resolved
resource files to be linked to their sources. A value of `copy` causes
resolved resource files to be copied.

For a period, the default will remain `link`. However, we plan to
change the default to `copy` in a future release.