Don't traverse symlinks when linking to run files for TensorBoard
It's not clear why symlinks were being traversed in the first
place. TensorBoard should avoid letting linked files in anyway. This
may have been a simple oversight.

If this change presents issues we need to make the official position
clear with proper tests (note that no tests broke with this change).