Stop-gap to address nagging test failures
The is portends a proper fix, which uses one thread to read both
stderr and stdout streams. As stderr is not buffered it should be
processed before stdout. The current scheme of two threads cannot
reliably synchronize the streams. This leads to spurious test
failures. The problem does not tend to manifest outside of tests.