Support for passing list values to operation
Guild uses YAML list syntax to denote batch arguments. E.g. the
following command triggers three runs:

    guild run op x=[1,2,3]

To pass the string value '[1,2,3]' requires shell quoting and also
requires that the operation script know to parse the value as a string
to generate a list. This runs contrary to Guild's "no code change"
claims.

This commit introduces a new flag attribute `arg-split`, which
indicates that a flag value should be split into parts to form a
list. This list value is applied to the operation according to the
flags interface, or args-dest attribute.

The value for `args-split` may be 'yes' or 'shlex' to indicate that
the flag value should be split using shell lexical syntax. The value
may alternatively be a string that is used as a token to split the
value - e.g. a comma, colon, etc.

    op:
      flags:
        x:
          arg-split: yes

To run op with multiple values, use shell-syntax:

    guild run op x="1 2 3"

For command line argument interfaces (e.g. argparse) Guild includes
split parts as command arguments for the applicable command line
option:

    python -m op --x 1 2 3

For globals interface Guild splits the value and applies YAML
formatting to each item to decode the part to a value.

See guild/tests/flags-dest.md#splitting-flag-args for examples.

Ref #226