Don't use LOCK with value 0 for terminated remote runs
LOCK should be only used for local processes and never considered for
remote runs. If the remote exit code is incorrect and we otherwise
know that the status is terminated, we need another signal to convey
that. This state begs the question, why is remote signalling
terminated without an exit status attr indicating terminated?