Delete default and random batch runs on successful completion
This is a potentially disruptive change. By default, default and
random batch runs are deleted. This commit includes generalized
support to delete runs on completion. This is useful for operations
that strictly perform work and do not need to be retained as a record.

Prior to this commit, when running a batch like this:

    $ guild run train.py x=[1,2]

Guild retained the batch run `train.py+`. This command therefore
generated three runs, not two. These batch runs serve little purpose
other than to indicate that a batch was run. They retain links to
their generated trials. Otherwise, they serve little purpose.

This commit includes two new options to the `run` command:

    --keep-batch  Use to prevent auto-delete of these batches
    --keep-run    Use to prevent auto-delete of normal runs

Both options are silently ignored if the applicable run or batch are
not otherwise auto-deleted.