Use pending rather than staged for initial trial status
It's incorrect to call initialized trials 'staged' - they aren't
staged but rather simply a copy of the proto run. Staging means that a
run's resources are resolved. This occurs when --stage-trials is used
for a batch. Otherwise an initialized trial is pending.

This has the advantage of avoiding races between batches and
queues/schedulers, though this is not the driver for this change.

Note that a pending run can be stared using `guild run --start <id>`
so the CLI to start trials after a terminated or failed batch is the
same.