Finagle regex support for operation name filters
Starting an operation filter term with '^' or ending one with '$'
causes Guild to treat the term as a regular expression.

This is a bit of a muddle. The alternatives:

- Continue without a deficient feature (e.g. can't distinguish between
  trials and optimizers, since they both contain the same operation
  name)

- Upgrade the term to a regex all the time. Dots, which are common in
  operation names, are now "any char".

- Define a special prefix that designates the string as a regex. This
  is basically what we've done here. We've also included a suffix.

- Introduce a new option - e.g. --label-regex, which users have to
  know to use.

Of these options, the commit maintains backward compatibility and
avoids a new option. The prefix '^' is reasonable in any case. The
suffix '$' is an added convenience.