Drop test support for Python 2.7 and 3.5
We're running into a wave of upstream deps that are dropping support
for Python 2.7, which is our sign/justification to drop it as
well. Plan to drop formally in 0.8. For now, we're just not testing as
the overhead to deal with the two lineages is too costly in light of
the pending dropped support.