Avoid initial sort of files data-table
Guild uses a natural sort algorithm server side to sort files by
name. Vuetify data-table uses standard lexical ordering. To support a
Guild defined sort order, we initially don't sort in the data
table. We also relax the requirement that the cols always sort. This
is not idea as we now support two sorts in the view - one from Guild
and another from Vuetify. However, this does at least provide the
option of a natural sort order in files.

Ref: #284