Save built view webapp in source
This is a fallback to use in cases where npm cannot build view for
whatever reason. Driver is a series of builds on AppVeyor for
0.7.4 (e.g. see https://bit.ly/30qnjfm)

To skip npm builds, set env var `SKIP_NPM=1` when running `setup.py
build|bdist_wheel`. This should not be needed if npm is otherwise
building successfully.