Rename `flags()` to `guild_flags()`
Ref: #311 #351

The core issue here is Guild's extension of Pandas. While this is
[officially supported by
Pandas](https://pandas.pydata.org/docs/development/extending.html), it
runs the risk of collision any time the Pandas API is changed.

We're emulating the Pandas API for runs and related structures to be
as Jupyter Notebook friendly as possible (note the `guild.ipy`
interface is explicitly designed for interactive use in a Jupyter
notebook). So when the upstrem Pandas API changes, the Guild interface
changes.

In this commit, we rename `flags()` to avoid redefining a new
attribute (as of 1.2) in the Pandas API.

Fundamentally, Guild needs [an official Python
API](https://github.com/guildai/guildai/issues/382) that isn't subject
breakage when another vendor modifies an API.

In the meantime, `guild.ipy` now works with current versions of
Pandas, but at the expense of a backward incompatible change.