Revert use of operation source code specs in package
This broke packaging (without spending too much time debugging - there
were None values included in the patterns list). I'm not sure we need
to worry too much about sync'ing package source with operation
source. Guild packaging uses the standard find_package to build a list
of included source files, which generally works as expected.

The source code specs in operation are not the smae pattern type as
what's expected in the package manifest.

We might consider dropping the use of find_packages and fully resolve
the list of operation source code files and include each file
explicitly in data files. I don't know if this is a hack of the Python
package spec, as data files are really non-source code files.