Avoid use of '_' for dummy vars and args
Motivation here is to encourage a bit more clarity even in cases where
the code seems obvious. While it's a common idiom in Python, I added a
lint rule to discourage this. In this refactor I found myself having
to answer the question, 'What is this thing I'm ignoring?' and finally
putting a name to it I think improves readability.