Use `click_util.completion_dir` only as a helper function
This is consistent with the other functions in `click_util`, which are
used by the shell completion callbacks - not as the callbacks
directly.

While there's a violation of DRY here (various duplications of
`_ac_dir` in the CLI interface modules) it avoids overloading
`completion_dir` as a dual purpose function - one that serves as the
callback and as the helper (used by callbacks. The duplication does
serve to insulate the logic of each param completion in that module -
a chage to one won't impact the others. If we want to avoid this
duplication, we could create another function in `click_util` that was
strictly a callback (i.e. conforms to the Click completion callback
interface).