Show applicable file completions for opspecs when ops are defined
There's some logic to NOT showing local files when there's a Guild
file that defines operations - the presumption is that the user has
bothered to define operations and only those should be shown rather
than clutter the list with local files (and all directories!)

However, this is a bit clever and comes across as surprising in
practice - why all of a sudden are local file completions missing?
Furthermore, local file completions start working when you specify a
partial. The behavior is quite subtle and looks like a bug, short of a
careful explanation.

In the end, let's keep things consistent and unclever here.