Cleanup config of Guild home in tests
Fixes a regression introduced in 10287ce0.

While we'd rather not be setting global state via
`set_guild_home()` (as well as `cd()`) in tests, this is a prevalent
pattern. This commit applies that pattern consistently. It removes the
`guild_home` kw arg to `run()` and instead relies on the interace
`guild.config.set_guild_home()` to init Guild home for tests.

Note this pattern works fine as long as tests using it are run in
series within a single Guild process. If/when we move to run tests in
parallel, the tests ought to be run in their own process.