Add '--force-uat' option to run uat from outside a venv
The default behavior for uat is to require a venv. This ensures that
any packages installed by uat don't accidentally modify a non-isolated
env. However, it can be useful to run from outside an activated
env (e.g. when running guild from ENV/bin/guild). In this case
'--force-uat' can be used to skip this check.