Don't use --user switch for package installs
Driver is a fix to CI tests. Guild is moving away from package support
so this is not a carefully considered decision. If there are
downstream issues caused by this change, we need to consider when the
process is run with a guild or python executable located in a virtual
env - i.e. modify `running_under_virtualenv()` to check for this.