Explicitly apply the config flags-dest setting to R opdef
The explanation for this is a bit pedantic, but here it is... When a
plugin provides an opdef, that's the end of it - no other plugins get
a chance to operation on that result. Plugins *do* cooperate in the
transform of *data* loaded (e.g. from a Guild file). But that's not
what's happening here - the plugin is providing the fully resolved
opdef. And so it must summon other plugins to apply their logic as
needed.

This is certainly not ideal as it requires plugins to know about other
plugins and apply them appropriately. Further, this is a *core*
feature that's implemented in a plugin. So it's all in a bit of a
hacked state.

Without thinking too much about it, future changes to Guild might
include any or all of the following:

1. Move the config file and flags-dest support into core

2. All plugins to cooperatively participate in the generation of an
opdef - this might include an explicit API to support new forms of
`flags-dest`

3. Provide a plugin developer tool check to inspect an opdef for
certain levels of 'completeness' - or at least provide a report as to
whaqt's there and what's not that might be there

Option 1 is probably not the right course, but rather option 2 so that
we can allow plugins to easily extend flags-dest - and then ensure
that flags-dest is properly supported for any opdef.

Option 3 is a Good Idea in general as it's frustrating to build
plugins in Guild without *any* solid idea that the generated opdefs
are correct, short of extensive testing.