Don't create github release on PyPI release
I'm not sure how we want to implement publishing releases to GitHub,
but we should address these issues when we do:

1. Tagging a commit *before* running CI tests to verify that the
commit is the release seems to be putting the cart before the horse. I
would imagine a scheme where we proposed a tag and that tag was added
by the CI only when all of the tests on all of the platforms pass.

2. Releases to PyPI seem straight forward and working, provided we
publish pre-releases to the production server (these should not be
published to the test server). But do we want to mirror these files on
GitHub as well? Who uses these? Considering you can get the same files
from PyPI if needed, what do GitHub releases get us? They would seem
to publish what's available but are there are other compelling reasons
to generate them?