Remove separate release workflow
Release is now baked into a single build/test/release workflow and
controlled by a user-selected workflow option.

We're not automatically generating releases when a release tag is
generated. I'm not sure what course is better, but tagging something
up front before it's cleared tests seems like the wrong order.

But in either case, generating a release is still a matter of taking a
manual step, be it tagging or some other manual trigger. The release
scheme as of this commit is to go to GitHub actions and manually start
the Build workflow with a checked option to generate a release. This
step does not tag anything or generate any GitHub releases - it simply
builds, tests, and then uploads the generated distributions (wheel and
source) to PyPI.

We could modify this to accept a 'release_tag' text workflow input,
rather than a 'release' boolean input. If specified, the workflow
would tag the commit, upload to PyPI and generate a GitHub release -
of course only if all of the CI tests pass.

This, or some other scheme entirely, is something we can look at down
the road to streamline the process.