Remove surprising use of globals as '__main__' module
Driver is https://my.guild.ai/t/torch-multiprocessing-spawn-fails/929

It appears that the original code effected the way help text was
displayed (see modified test in this commit) but it substituted a
globals dict for a module, which broke the Python env. If there's a
need to format help text differently, let's revisit this
implementation.